+++ Sources/PeripheryKit/Indexer/Declaration.swift
@@ -177,6 +177,7 @@ public final class Declaration {
     public var declaredType: String?
     public var letShorthandIdentifiers: Set<String> = []
     public var hasCapitalSelfFunctionCall: Bool = false
+    public var hasGenericFunctionReturnedMetatypeParameters: Bool = false
 
     public var parent: Declaration?
     var related: Set<Reference> = []