+++ Tests/AccessibilityTests/AccessibilityProject/Sources/TargetA/PublicTypeUsedAsPublicPropertyType.swift
@@ -1,12 +1,21 @@
 import Foundation
 
-public class PublicTypeUsedAsPublicPropertyType {}
+public class PublicTypeUsedAsPublicPropertyType1 {}
+public class PublicTypeUsedAsPublicPropertyType2 {}
+public class PublicTypeUsedAsPublicPropertyType3 {}
+public class PublicTypeUsedAsPublicPropertyType4 {}
 public struct PublicTypeUsedAsPublicPropertyGenericArgumentType: Hashable {}
 public class PublicTypeUsedAsPublicPropertyArrayType {}
 
 public class PublicTypeUsedAsPublicPropertyTypeRetainer {
     public init() {}
-    public var retain1: PublicTypeUsedAsPublicPropertyType?
+
+    public var retain: [Any?] {
+        [retain1, retain2, retain3, retain4, retain5, retain6]
+    }
+
+    public var retain1: PublicTypeUsedAsPublicPropertyType1?
     public var retain2: Set<PublicTypeUsedAsPublicPropertyGenericArgumentType>?
     public var retain3: [PublicTypeUsedAsPublicPropertyArrayType] = []
+    public var (retain4, (retain5, retain6)): (PublicTypeUsedAsPublicPropertyType2, (PublicTypeUsedAsPublicPropertyType3, PublicTypeUsedAsPublicPropertyType4)) = (.init(), (.init(), .init()))
 }